Revert "Development movie card front end"
