--- media-wrangler-back-end/src/main/java/com/mediawrangler/media_wrangler/dto/RegisterRequest.java
@@ -1,50 +1,0 @@
-package com.mediawrangler.media_wrangler.dto;
-
-public class RegisterRequest {
-    private String username;
-    private String firstName;
-    private String lastName;
-    private String password;
-    private String email;
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-}
