+++ media-wrangler-back-end/src/main/java/com/mediawrangler/media_wrangler/models/User.java
@@ -33,18 +33,6 @@ public class User {
     @Email(message = "Invalid email format")
     private String email;
 
-    public User(String username, String firstname, String lastname, String password, String email) {
-        this.username = username;
-        this.firstname = firstname;
-        this.lastname = lastname;
-        this.password = password;
-        this.email = email;
-    }
-
-    public User() {
-    }
-
-
     public int getId() {
         return id;
     }
