+++ media-wrangler-front-end/package.json
@@ -18,10 +18,7 @@
     "bulma": "^1.0.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^7.1.1",
-    "react-icons": "^5.4.0",
-    "react-star-ratings": "^2.3.0"
-
+    "react-router-dom": "^7.0.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.15.0",
