+++ media-wrangler-front-end/src/App.css
@@ -1,31 +1,5 @@
-<<<<<<< HEAD
 /* .App {
     text-align: center;
-=======
-/* #root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
-}
-
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
-}
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-.logo.react:hover {
-  filter: drop-shadow(0 0 2em #61dafbaa);
-}
-
-@keyframes logo-spin {
-  from {
-    transform: rotate(0deg);
->>>>>>> movie-card-front-end
   }
   
   .App-logo {
@@ -38,7 +12,6 @@
       animation: App-logo-spin infinite 20s linear;
     }
   }
-<<<<<<< HEAD
   
   .App-header {
     background-color: #282c34;
@@ -64,14 +37,3 @@
     }
   }
    */
-=======
-}
-
-.card {
-  padding: 2em;
-}
-
-.read-the-docs {
-  color: #888;
-} */
->>>>>>> movie-card-front-end
