+++ media-wrangler-front-end/src/Services/RegistrationService.js
@@ -3,10 +3,8 @@ import axios from "axios";
 export async function apiRegister(registerData) {
     try {
         const response = await axios.post(
-            'http://localhost:8080/users/register',
-            registerData, {
-                withCredentials: true,
-            }
+            "http://localhost:8080/register",
+            registerData
         );
         console.log("Response:", response);
         if (response.status === 201) {
