--- media-wrangler-front-end/src/components/MoviePosterCard/MovieCard.css
@@ -1,26 +1,0 @@
-/* Movie Card for Search Results and Lists */
-
-
-#movie-search {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(70px, 1fr)); /* Create flexible grid columns */
-    gap: 10px; /* Add spacing between grid items */
-    padding: 10px;
-    justify-items: center; /* Optional: center the images within their grid cells */
-}
-
-#movie-search div {
-    display: flex; /* Make the div a flex container */
-    justify-content: center; /* Center image horizontally */
-    align-items: center; /* Center image vertically */
-}
-
-#movie-search img {
-    width: 100%; /* Ensure the image fills the grid item */
-    height: auto; /* Maintain aspect ratio */
-    border-radius: 5px; /* Optional: add rounded corners to the image */
-}
-
-
-
-/* Movie Card for ... */
