--- media-wrangler-front-end/src/components/Rating/StarRating.css
@@ -1,5 +1,0 @@
-input[type=radio] {
-    position: fixed;
-    opacity: 0;
-    pointer-events: none;
-}
