+++ media-wrangler-front-end/src/components/Register/Register.jsx
@@ -1,4 +1,3 @@
-import axios from "axios";
 import React, { useState } from "react"
 import { useNavigate } from "react-router-dom";
 import { apiRegister } from "../../Services/RegistrationService";
@@ -28,7 +27,7 @@ export default function Register() {
 
         let responseMessage = await apiRegister(registerData);
 
-         console.log(responseMessage);
+        console.log(responseMessage);
         if (responseMessage === "Success") {
             navigate("/login");
         } else {
