--- media-wrangler-front-end/src/components/ReviewForm/BasicReviewForm.css
@@ -1,39 +1,0 @@
-
- 
- /*MovieReview Component */
-
- .create-review {
-    max-width: 800px;
-    margin: 0 auto;
-    text-align: center;
-    padding: 20px;
-}
-
-.create-review label {
-    text-align: left;
-    display: block;
-    color: #544EFD;
-}
-
-.create-review h3 {
-  font-size: 20px;
-  color: purple;
-  margin-bottom: 30px;
-}
-
-.create-review input,.create-review textarea, .create-review select {
-  width: 100%;
-  padding: 6px 10px;
-  margin: 10px 0;
-  border: 2px solid #544EFD;
-  border-radius: 8px;
-  background-color: black;
-}
-
-
-.create-review button {
-  background: green;
-  border: 0px;
-  border-radius: 8px;
-  cursor: pointer;
-}
