--- media-wrangler-front-end/src/components/ReviewForm/ReviewForm.css
@@ -1,34 +1,0 @@
-/* 
-  I just googled some styling to get a visual going, but we can adjust once everything is functional and we are beautifying things
-  I think that I need to rename the css files that are for the specific components. I remember something about adding module into, maybe?
-  Helps with scope issues if I am correct. Once we get ready to style I can go back through and update it all 
-*/
-
-
-.review-container {
-  display: flex;
-  justify-content: space-between; 
-  align-items: center; 
-  gap: 20px; 
-}
-
-.poster {
-  flex: 0 0 30%; 
-  max-width: 600px; 
-  max-height: 700px;
-}
-
-.poster img {
-  width: 100%; 
-  height: auto; 
-  border-radius: 8px; 
-}
-
-.review-form {
-  flex: 1; 
-  max-width: 600px; 
-}
-
-.checkbox {
-  margin-left: 10px; /* Adds space between the two radio buttons */
-}
